% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adj2igraph}
\alias{adj2igraph}
\title{Convert adjacency matrix to directed igraph with edge attributes}
\usage{
adj2igraph(
  G,
  pMax,
  chisqMin,
  absPcorMin,
  Threshold,
  sampleSize
)
}
\arguments{
\item{G}{Adjacency matrix.}

\item{pMax}{P-value matrix corresponding to edges in `G`.}

\item{chisqMin}{Chi-square statistics matrix corresponding to edges in `G`.}

\item{absPcorMin}{Partial correlation matrix corresponding to edges in `G`.}

\item{Threshold}{Threshold matrix corresponding to edges in `G`.}

\item{sampleSize}{sampleSize matrix corresponding to edges in `G`.}
}
\value{
Directed igraph object with edge attributes `pMax` and `chisqMin`.
}
\description{
Constructs a directed igraph object from an adjacency matrix, assigns `pMax` and `chisqMin` values to each edge.
}
