% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_prediction.R
\name{apply_expression_model}
\alias{apply_expression_model}
\title{Apply Expression Model}
\usage{
apply_expression_model(models, Y, group, ncores)
}
\arguments{
\item{models}{List of fitted models (See \code{\link{fit_expression_model}}).}

\item{Y}{Expression matrix (cells x genes).}

\item{group}{Vector indicating cell groups.}

\item{ncores}{Number of cores for parallel computation.}
}
\value{
Predicted expression matrix (cells x genes).
}
\description{
Predict gene expression for test cells using fitted models.
}
