% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_prediction.R
\name{apply_perturbation}
\alias{apply_perturbation}
\title{Apply Perturbation}
\usage{
apply_perturbation(
  ko_expressions,
  wt_expressions,
  graph,
  models,
  max_dist = Inf
)
}
\arguments{
\item{ko_expressions}{Known KO gene expressions.}

\item{wt_expressions}{WT bulk gene expressions.}

\item{graph}{Graph structure (igraph object or 'all').}

\item{models}{List of fitted models (See \code{\link{fit_expression_model}}).}

\item{max_dist}{Maximum distance to propagate perturbation effect (default is Inf).}
}
\value{
Predicted expression matrix (KO genes x all genes).
}
\description{
Simulate gene expression after perturbation using graph structure.
}
