% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_auc.R
\name{calc_auc}
\alias{calc_auc}
\title{Calculate AUC measuring consistency between correlation and differential expression}
\usage{
calc_auc(stat, thr, ncores = 1)
}
\arguments{
\item{stat}{DE statistics, see \code{\link{calc_perturbation_effect}}.}

\item{thr}{Numeric value in [0, 1) indicating the minimum quantile of absolute correlation to include in AUC calculation.}

\item{ncores}{Number of cores (default is 1).}
}
\value{
AUC matrix with rows representing KOs and two AUC columns based on 'pearson' and 'spearman' correlations.
}
\description{
Calculate AUC measuring consistency between Pearson/Spearman correlations and Cohen' D, under specified threshold.
}
\examples{
# x -> y -> z
set.seed(123)
n <- 1e3
x0 <- rnorm(n)
x1 <- rnorm(n, mean = -1)
x <- c(x0, x1)
group <- c(rep('WT', n), rep('x', n))
y <- rnorm(2 * n, mean = x, sd = 1)
z <- rnorm(2 * n, mean = y, sd = 1)
data <- cbind(x = x, y = y, z = z)
rownames(data) <- paste0("cell", seq_len(nrow(data)))
names(group) <- rownames(data)
stat <- calc_perturbation_effect(
  Y = data,
  group = group,
  ncores = 2
)
calc_auc(stat, thr = 0)
}
