% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_correlation.R
\name{calc_count_cdf}
\alias{calc_count_cdf}
\title{Proportion of counts ≤ threshold}
\usage{
calc_count_cdf(count, ncores, max_thr = 10)
}
\arguments{
\item{count}{scRNA-seq count matrix (cells × genes).}

\item{ncores}{Number of cores for parallelization.}

\item{max_thr}{Maximum threshold (default: 10).}
}
\value{
Matrix (thresholds × genes) of proportions.
}
\description{
Compute proportion of cells with counts ≤ thresholds for each gene.
}
\examples{
count <- matrix(sample(0:15, 50, replace = TRUE), nrow = 10)
dimnames(count) <- list(paste0('cell', seq(10)), paste0('gene', seq(5)))
calc_count_cdf(count, ncores = 1, max_thr = 10)
}
