% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_correlation.R
\name{calc_pcor}
\alias{calc_pcor}
\title{Calculate partial correlation for sparse count data}
\usage{
calc_pcor(i, j, k, count, Y, max_thr, min_n1, min_n2)
}
\arguments{
\item{i}{Integer.}

\item{j}{Integer.}

\item{k}{Integer representing index of conditional variable.}

\item{count}{Count matrix (cells x genes).}

\item{Y}{Normalized expression matrix (cells x genes).}

\item{max_thr}{Maximum threshold for conditional variable.}

\item{min_n1}{Minimum number of samples satisfying Yk > selected threshold.}

\item{min_n2}{Minimum number of samples satisfying Yk > selected threshold, Yi > 0, and Yj > 0.}
}
\value{
Data frame of threshold, sample size, partial correlation estimate, z-score, and p-value.
}
\description{
Correct partial correlation for sparse count data by thresholding conditional variable.
}
\examples{
# Gene 1 -> Gene 2 -> Gene 3
set.seed(123)
n <- 1e5
u1 <- rnorm(n)
u2 <- rnorm(n, u1)
u3 <- rnorm(n, u2)
u <- cbind(g1 = u1, g2 = u2, g3 = u3)
count <- apply(u, 2, function(x) rpois(n, lambda = exp(x)))
Y <- log1p(count)
# True partial correlation in latent space
ppcor::pcor(u)
# Uncorrected partial correlation
ppcor::pcor.test(Y[, 1], Y[, 3], Y[, 2])
ppcor::pcor.test(Y[, 1], Y[, 2], Y[, 3])
# Corrected partial correlation
calc_pcor(i = 1, j = 3, k = 2, count = count, Y = Y, max_thr = 20, min_n1 = 1000, min_n2 = 200)
calc_pcor(i = 1, j = 2, k = 3, count = count, Y = Y, max_thr = 20, min_n1 = 1000, min_n2 = 200)
}
