% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_graph_with_gp}
\alias{collapse_graph_with_gp}
\title{Collapse graph with gene programs}
\usage{
collapse_graph_with_gp(graph, pname, pgenes, Y, allow_pgene_out)
}
\arguments{
\item{graph}{igraph object.}

\item{pname}{Gene program name.}

\item{pgenes}{Gene program genes.}

\item{Y}{scRNA-seq normalized expression matrix.}

\item{allow_pgene_out}{Allow program genes as parents of non-program genes or not.}
}
\value{
igraph object.
}
\description{
Convert g -> program to g -> all program genes, and drop program.
}
