% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_prediction_with_gp.R
\name{fit_expression_model_with_gp}
\alias{fit_expression_model_with_gp}
\title{Fit a Two-Stage Portability Prediction Model}
\usage{
fit_expression_model_with_gp(
  Y,
  group,
  graph,
  source_effects,
  pname,
  pgenes,
  ncores
)
}
\arguments{
\item{Y}{A numeric matrix of expression data for the target dataset.}

\item{group}{A character vector for the cell groups in Y.}

\item{graph}{An igraph object for the target dataset.}

\item{source_effects}{A named list of data frames with KO effects from source datasets.}

\item{pname}{The name of the hub node (e.g., "PC1").}

\item{pgenes}{A character vector of the signature genes.}

\item{ncores}{The number of cores for parallel computation.}
}
\value{
A list containing the `B_matrix` and a list of trained `portability_models`.
}
\description{
Trains two model components: a standard GRN model (B_matrix) for the target
dataset, and a flexible set of portability models to predict the SCOPE
effect using information from source datasets.
}
