% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sepset}
\alias{get_sepset}
\title{Get separation set of two genes}
\usage{
get_sepset(skel, g1, g2)
}
\arguments{
\item{skel}{See \code{\link{infer_skeleton}}.}

\item{g1}{Gene name.}

\item{g2}{Gene name.}
}
\value{
Separation set.
}
\description{
Extract separation set of two genes from graph skeleton
}
