% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_cgrn_with_gp.R
\name{infer_causalgrn_with_gp}
\alias{infer_causalgrn_with_gp}
\title{Direct edges using perturbation effect with gene program}
\usage{
infer_causalgrn_with_gp(
  graph,
  stat,
  alpha,
  pname,
  pgenes,
  allow_pgene_out,
  conservative = TRUE
)
}
\arguments{
\item{graph}{Initial igraph object.}

\item{stat}{Perturbation effect from \code{\link{calc_perturbation_effect}}.}

\item{alpha}{Numeric representing DE Q-value threshold.}

\item{pname}{Gene program name.}

\item{pgenes}{Gene program genes.}

\item{allow_pgene_out}{Allow program genes as parents of non-program genes or not.}

\item{conservative}{Logical indicating whether to make conservative inference (Default is \code{TRUE}).}
}
\value{
igraph object.
}
\description{
Infers causal GRN by directing edges based on differential expression caused by perturbations.
}
