% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_skeleton.R
\name{infer_skeleton}
\alias{infer_skeleton}
\title{Infer graph skeleton via conditional independence tests}
\usage{
infer_skeleton(
  count,
  Y,
  alpha,
  min_abspcor,
  ncores,
  G = NULL,
  max_order = 1,
  max_thr = 10,
  min_n1 = 1000,
  min_n2 = 200,
  sepset = TRUE
)
}
\arguments{
\item{count}{scRNA-seq count matrix (cells x genes).}

\item{Y}{scRNA-seq normalized expression matrix (cells × genes). For example, log1p(total UMI corrected count).}

\item{alpha}{Significance level for CI tests.}

\item{min_abspcor}{Minimum absolute value of partial correlation for kept edges.}

\item{ncores}{Number of CPU cores for parallel processing.}

\item{G}{Optional initial adjacency matrix (defaults to a fully connected graph without self-loops).}

\item{max_order}{Maximum conditioning set size (0 or 1, default is 1).}

\item{max_thr}{Maximum threshold for conditional variable (default is 10).}

\item{min_n1}{Minimum number of samples satisfying Yk > selected threshold (default is 1000).}

\item{min_n2}{Minimum number of samples satisfying Yk > selected threshold, Yi > 0, and Yj > 0 (default is 200).}

\item{sepset}{Return separation set or not (default is \code{TRUE}).}
}
\value{
A list with:
\describe{
  \item{graph}{An igraph object representing the inferred skeleton with edge attributes `pMax`, `chisqMin`, and `absPcorMin`.}
  \item{sepSet}{List of separation sets for each node pair.}
}
}
\description{
Constructs graph skeleton by performing iterative conditional independence (CI) tests on gene expression data.
Edges are removed based on statistical independence.
}
\examples{
# Gene 1 -> Gene 2 -> Gene 3
set.seed(123)
n <- 1e5
u1 <- rnorm(n)
u2 <- rnorm(n, u1)
u3 <- rnorm(n, u2)
u <- cbind(g1 = u1, g2 = u2, g3 = u3)
count <- apply(u, 2, function(x) rpois(n, lambda = exp(x)))
rownames(count) <- paste0("cell", seq_len(nrow(count)))
Y <- log1p(count)
skel <- infer_skeleton(
  count = count,
  Y = Y,
  alpha = 1e-3,
  min_abspcor = 0.05,
  ncores = 2,
  max_order = 1
)
print(igraph::as_data_frame(skel$graph))
plot(skel$graph)
get_sepset(skel, 'g1', 'g3')

}
