% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_skeleton_from_perturbed.R
\name{infer_skeleton_from_perturbed}
\alias{infer_skeleton_from_perturbed}
\title{Infer graph skeleton via conditional independence tests based on wild-type & perturbed data}
\usage{
infer_skeleton_from_perturbed(
  Y,
  group,
  alpha,
  ncores,
  G = NULL,
  max_order = 1,
  hub_genes = NULL,
  max_order_hub = NULL,
  sepset = TRUE
)
}
\arguments{
\item{Y}{scRNA-seq matrix (samples × genes).}

\item{group}{Named vector of cell label: 'WT' for wild-type cells, perturbed gene for perturbed cell.}

\item{alpha}{Significance level for CI tests.}

\item{ncores}{Number of CPU cores for parallel processing.}

\item{G}{Optional initial adjacency matrix; defaults to a fully connected graph without self-loops.}

\item{max_order}{Maximum conditioning set size for non-hub genes.}

\item{hub_genes}{Optional character vector of hub gene names.}

\item{max_order_hub}{Maximum conditioning set size for hub genes.}

\item{sepset}{Return separation set or not (default is \code{TRUE}).}
}
\value{
A list with:
\describe{
  \item{graph}{An igraph object representing the inferred skeleton with edge attributes `pMax` and `chisqMin`.}
  \item{sepSet}{List of separation sets for each node pair.}
}
}
\description{
Constructs graph skeleton by performing iterative conditional independence (CI) tests on gene expression data.
Edges are removed based on statistical independence, optionally considering hub genes with higher-order conditioning.
}
\examples{
# No hub genes
# x -> y -> z
set.seed(123)
n <- 1e3
x0 <- rnorm(n)
x1 <- rnorm(n, mean = -1)
x <- c(x0, x1)
group <- c(rep('WT', n), rep('x', n))
y <- rnorm(2 * n, mean = x, sd = 1)
z <- rnorm(2 * n, mean = y, sd = 1)
data <- cbind(x = x, y = y, z = z)
skel <- infer_skeleton_from_perturbed(
  Y = data,
  group = group,
  alpha = 1e-3,
  ncores = 2,
  max_order = 1
)
print(igraph::as_data_frame(skel$graph))
plot(skel$graph)
get_sepset(skel, 'x', 'z')

# With hub genes
# Hub genes PC 1-3, x_i -> PC_j -> z
set.seed(123)
n <- 1e3
X0 <- matrix(rnorm(5 * n), n, 5)
X1 <- cbind(rnorm(n, mean = -1), matrix(rnorm(4 * n, mean = -1), n, 4))
X <- rbind(X0, X1)
group = c(rep('WT', n), rep('x1', n))
B <- cbind(
  c(0.4082, 0.4082, 0.4082, 0.4082, 0.4082),
  c(0.6325, 0.3162, 0.0000, -0.3162, -0.6325),
  c(0.0000, 0.7071, -0.7071, 0.0000, 0.0000)
)
PC <- X \%*\% B + matrix(rnorm(3 * 2 * n), 2 * n, 3)
print(cor(PC))
z <- PC \%*\% rep(1, 3) + rnorm(2 * n)
data <- cbind(X, PC, z)
colnames(data) <- c(paste0('x', 1:5), paste0('PC', 1:3), 'z')
G <- matrix(TRUE, ncol(data), ncol(data), dimnames = list(colnames(data), colnames(data)))
diag(G) <- FALSE
G[paste0('PC', 1:3), paste0('PC', 1:3)] <- FALSE
skel <- infer_skeleton_from_perturbed(
  Y = data,
  group = group,
  alpha = 1e-6,
  ncores = 2,
  max_order = 1,
  hub_genes = paste0('PC', 1:3),
  max_order_hub = 3
)
plot(skel$graph, layout = igraph::layout_with_kk)
}
