% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_skeleton_from_wildtype.R
\name{infer_skeleton_from_wildtype}
\alias{infer_skeleton_from_wildtype}
\title{Infer graph skeleton via conditional independence tests based on wild-type data}
\usage{
infer_skeleton_from_wildtype(
  Y,
  alpha,
  ncores,
  G = NULL,
  max_order = 1,
  hub_genes = NULL,
  max_order_hub = NULL
)
}
\arguments{
\item{Y}{scRNA-seq matrix (samples × genes).}

\item{alpha}{Significance level for CI tests.}

\item{ncores}{Number of CPU cores for parallel processing.}

\item{G}{Optional initial adjacency matrix; defaults to a fully connected graph without self-loops.}

\item{max_order}{Maximum conditioning set size for non-hub genes.}

\item{hub_genes}{Optional character vector of hub gene names.}

\item{max_order_hub}{Maximum conditioning set size for hub genes.}
}
\value{
A list with:
\describe{
  \item{graph}{An igraph object representing the inferred skeleton with edge attributes `pMax` and `chisqMin`.}
  \item{sepSet}{List of separation sets for each node pair.}
}
}
\description{
Constructs graph skeleton by performing iterative conditional independence (CI) tests on gene expression data.
Edges are removed based on statistical independence, optionally considering hub genes with higher-order conditioning.
}
\examples{
# No hub genes
# x -> y -> z
set.seed(123)
n <- 1e3
x <- rnorm(n)
y <- rnorm(n, mean = x, sd = 1)
z <- rnorm(n, mean = y, sd = 1)
data <- cbind(x = x, y = y, z = z)
res <- infer_skeleton_from_wildtype(
  Y = data,
  alpha = 0.05,
  ncores = 2,
  max_order = 1
)
print(igraph::as_data_frame(res$graph))
plot(res$graph)
print(res$sepSet[[1]][[3]])

# With hub genes
# Hub genes PC 1-3, x_i -> PC_j -> z
set.seed(123)
n <- 1e3
X <- matrix(rnorm(5 * n), n, 5)
B <- cbind(
  c(0.4082, 0.4082, 0.4082, 0.4082, 0.4082),
  c(0.6325, 0.3162, 0.0000, -0.3162, -0.6325),
  c(0.0000, 0.7071, -0.7071, 0.0000, 0.0000)
)
PC <- X \%*\% B + matrix(rnorm(3 * n), n, 3)
print(cor(PC))
z <- PC \%*\% rep(1, 3) + rnorm(n)
data <- cbind(X, PC, z)
colnames(data) <- c(paste0('x', 1:5), paste0('PC', 1:3), 'z')
G <- matrix(TRUE, ncol(data), ncol(data), dimnames = list(colnames(data), colnames(data)))
diag(G) <- FALSE
G[paste0('PC', 1:3), paste0('PC', 1:3)] <- FALSE
res <- infer_skeleton_from_wildtype(
  Y = data,
  alpha = 1e-3,
  ncores = 2,
  max_order = 1,
  hub_genes = paste0('PC', 1:3),
  max_order_hub = 3
)
plot(res$graph, layout = igraph::layout_with_kk)
}
