% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parallel_cor}
\alias{parallel_cor}
\title{Compute Pearson correlation matrix in parallel}
\usage{
parallel_cor(mat, ncores)
}
\arguments{
\item{mat}{Matrix.}

\item{ncores}{Number of cores.}
}
\value{
Pearson correlation matrix.
}
\description{
Efficiently computes the correlation matrix of a numeric matrix, using multiple cores for large matrices.
}
