% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_skeleton.R
\name{perform_ci_test}
\alias{perform_ci_test}
\title{Perform conditional independence tests on graph edges}
\usage{
perform_ci_test(
  G,
  order,
  max_order,
  count,
  Y,
  max_thr,
  min_n1,
  min_n2,
  alpha,
  min_abspcor,
  pMax,
  chisqMin,
  absPcorMin,
  Threshold,
  sampleSize,
  sepSet,
  ncores
)
}
\arguments{
\item{G}{Adjacency matrix.}

\item{order}{Integer specifying the size of the conditioning set.}

\item{max_order}{Maximum allowed size for the conditioning set.}

\item{count}{scRNA-seq count matrix (cells x genes).}

\item{Y}{scRNA-seq normalized expression matrix (cells × genes). For example, log1p(total UMI corrected count).}

\item{max_thr}{Maximum threshold for conditional variable.}

\item{min_n1}{Minimum number of samples satisfying Yk > selected threshold.}

\item{min_n2}{Minimum number of samples satisfying Yk > selected threshold, Yi > 0, and Yj > 0.}

\item{alpha}{Significance level for the CI tests.}

\item{min_abspcor}{Minimum absolute value of partial correlation for kept edges.}

\item{pMax}{Matrix to store the maximum p-values for each edge.}

\item{chisqMin}{Matrix to store the minimum chi-square statistics for each edge.}

\item{absPcorMin}{Matrix to store the minimum absolute value of partial correlation for each edge.}

\item{Threshold}{Matrix to store the count threshold of conditional variable when absPcorMin is achieved.}

\item{sampleSize}{Matrix to store the sample size of the test giving absPcorMin.}

\item{sepSet}{List of separation sets for each pair of nodes.}

\item{ncores}{Number of cores to use for parallel processing.}
}
\value{
A list containing:
\describe{
  \item{G}{Updated adjacency matrix after removing conditionally independent edges.}
  \item{pMax}{Updated matrix of maximum p-values for each edge.}
  \item{chisqMin}{Updated matrix of minimum chi-square statistics for each edge.}
  \item{absPcorMin}{Updated matrix of minimum absolute value of partial correlation for each edge.}
  \item{Threshold}{Updated matrix of count threshold of conditional variable for each edge.}
  \item{sampleSize}{Updated matrix of sample size of test for each edge.}
  \item{sepSet}{Updated list of separation sets for each pair of nodes.}
  \item{done}{Logical indicating whether all edges reached the specified order.}
}
}
\description{
Applies conditional independence (CI) tests to each edge in the graph `G`. Edges are evaluated in parallel,
and those found conditionally independent given a subset of variables are removed.
The function updates the graph and associated statistics accordingly.
}
