% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_skeleton_from_perturbed.R
\name{perform_ci_test_from_perturbed}
\alias{perform_ci_test_from_perturbed}
\title{Perform conditional independence tests on graph edges based on wild-type & perturbed data}
\usage{
perform_ci_test_from_perturbed(
  G,
  order,
  max_order,
  hub_index,
  Cs,
  ns,
  ko_idx,
  alpha,
  pMax,
  chisqMin,
  sepSet,
  ncores
)
}
\arguments{
\item{G}{Adjacency matrix.}

\item{order}{Integer specifying the size of the conditioning set.}

\item{max_order}{Maximum allowed size for the conditioning set.}

\item{hub_index}{Vector of node indices considered as hubs.}

\item{Cs}{Correlation matrices of the variables within each group.}

\item{ns}{Sample sizes of each group.}

\item{ko_idx}{Index of KO genes.}

\item{alpha}{Significance level for the CI tests.}

\item{pMax}{Matrix to store the maximum p-values for each edge.}

\item{chisqMin}{Matrix to store the minimum chi-square statistics for each edge.}

\item{sepSet}{List of separation sets for each pair of nodes.}

\item{ncores}{Number of cores to use for parallel processing.}
}
\value{
A list containing:
\describe{
  \item{G}{Updated adjacency matrix after removing conditionally independent edges.}
  \item{pMax}{Updated matrix of maximum p-values for each edge.}
  \item{chisqMin}{Updated matrix of minimum chi-square statistics for each edge.}
  \item{sepSet}{Updated list of separation sets for each pair of nodes.}
  \item{done}{Logical indicating whether all edges reached the specified order.}
}
}
\description{
Applies conditional independence (CI) tests to each edge in the graph `G`. Edges are evaluated in parallel,
and those found conditionally independent given a subset of variables are removed.
The function updates the graph and associated statistics accordingly.
}
