% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_perturbation_effect.R
\name{plot_adj_pv_for_ko_pairs}
\alias{plot_adj_pv_for_ko_pairs}
\title{Plot adjusted P-values for KO gene pairs in -log10 scale}
\usage{
plot_adj_pv_for_ko_pairs(stat, eps = 1e-05)
}
\arguments{
\item{stat}{Tibble containing DE statistics (See \code{\link{calc_perturbation_effect}}).}

\item{eps}{Numeric, minimum value to clamp p-values to. Default is 1e-5.}
}
\value{
ggplot object.
}
\description{
Plot adjusted P-values for KO gene pairs in -log10 scale
}
