% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_perturbation_effect.R
\name{plot_cd_for_ko_pairs}
\alias{plot_cd_for_ko_pairs}
\title{Plot Cohen's D for KO gene pairs}
\usage{
plot_cd_for_ko_pairs(stat)
}
\arguments{
\item{stat}{Tibble containing DE statistics (See \code{\link{calc_perturbation_effect}}).}
}
\value{
ggplot object.
}
\description{
Plot Cohen's D for KO gene pairs
}
