% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines_prediction.R
\name{predict_oracle_perturbation_effect}
\alias{predict_oracle_perturbation_effect}
\title{Predict Oracle Perturbation Effect}
\usage{
predict_oracle_perturbation_effect(B, Y, group, wt_expressions)
}
\arguments{
\item{B}{A numeric matrix of regulatory coefficients where rows are sources
('Intercept' and genes) and columns are targets (genes). This is the output
of \code{\link{fit_expression_model}}.}

\item{Y}{A numeric matrix of expression data (cells x genes) for which to
make predictions.}

\item{group}{Named character vector indicating the group for each cell in
\code{Y}. Wild-type cells are not required for oracle prediction.}

\item{wt_expressions}{A named numeric vector of the wild-type expression levels
for each gene.}
}
\value{
A numeric matrix of predicted delta values (rows are unique KOs,
  columns are genes).
}
\description{
Predicts the perturbation effect (delta from wild-type) for test cells using
a fitted B matrix, where the true expression values of the predictors are
known.
}
