% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_prediction_with_gp.R
\name{predict_standard_effect_with_gp}
\alias{predict_standard_effect_with_gp}
\title{Predict Perturbation Effect Using a Two-Stage Portability Model}
\usage{
predict_standard_effect_with_gp(
  B,
  portability_models,
  source_effects,
  ko_expressions,
  wt_expressions,
  pname,
  max_dist = Inf,
  ncores = 1
)
}
\arguments{
\item{B}{A numeric matrix of regulatory coefficients.}

\item{portability_models}{A named list of trained `lm` objects.}

\item{source_effects}{A named list of data frames with predictor variables.}

\item{ko_expressions}{A named vector of expression levels for the test KOs.}

\item{wt_expressions}{A named vector of WT expressions for the target dataset.}

\item{pname}{The name of the hub node.}

\item{max_dist}{Maximum distance for effect propagation.}

\item{ncores}{Number of cores for parallel computation.}
}
\value{
A numeric matrix of predicted delta values for the test KOs.
}
\description{
Predict Perturbation Effect Using a Two-Stage Portability Model
}
