% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines_grn.R
\name{run_cor}
\alias{run_cor}
\title{Build correlation based GRN}
\usage{
run_cor(Y)
}
\arguments{
\item{Y}{Matrix of normalized scRNA-seq data.}
}
\value{
igraph object.
}
\description{
Build GRN based on marginal correlations between genes.
}
