% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines_grn.R
\name{run_genie3}
\alias{run_genie3}
\title{Run GENIE3 Algorithm}
\usage{
run_genie3(wt, ncores, verbose = FALSE)
}
\arguments{
\item{wt}{scRNA-seq matrix of wild-type cells.}

\item{ncores}{Number of cores.}

\item{verbose}{Whether to print training details (default is FALSE).}
}
\value{
igraph object.
}
\description{
Run GENIE3 algorithm and return an igraph object.
}
