% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines_grn.R
\name{run_ges}
\alias{run_ges}
\title{Run GES Algorithm}
\usage{
run_ges(wt, verbose = FALSE)
}
\arguments{
\item{wt}{scRNA-seq matrix of wild-type cells.}

\item{verbose}{Whether to print training details (default is FALSE).}
}
\value{
igraph object.
}
\description{
Run a GES algorithm and return an igraph object.
}
