% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_expression.R
\name{simulate_expression}
\alias{simulate_expression}
\title{GRN-guided simulation of scRNA-seq data of wild-type and perturbed cells mimicking real data}
\usage{
simulate_expression(d, Y, group, min_coef, max_coef, cv, covariates = NULL)
}
\arguments{
\item{d}{Expected number of neighbors (in/out).}

\item{Y}{Real single-cell perturbation data (cells x genes).}

\item{group}{Named vector of cell label: 'WT' for wild-type cells, perturbed gene for perturbed cell.}

\item{min_coef}{Minimum value of simulated coefficients.}

\item{max_coef}{Maximum value of simulated coefficients.}

\item{cv}{Coefficient of variation for simulated normal distributions.}

\item{covariates}{Cell covariates.}
}
\value{
List containing:
\itemize{
  \item \code{dag}: Simulated DAG.
  \item \code{coef}: Simulated coefficients.
  \item \code{Y}: Simulated expression.
  \item \code{group}: Simulated group.
}
}
\description{
Simulates gene expression based on simulated directed acyclic graph and real single-cell perturbation data.
}
