% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_difference.R
\name{structure_difference_at_breaks}
\alias{structure_difference_at_breaks}
\title{Evaluate Structure Difference at Top-Edge Breaks}
\usage{
structure_difference_at_breaks(pred, g0, ntops, ncores = 1)
}
\arguments{
\item{pred}{Data frame of predicted directed edges with columns \code{from},
\code{to}, and \code{score}.}

\item{g0}{igraph object of the true graph.}

\item{ntops}{Vector of number of top predicted edges to evaluate.}

\item{ncores}{Number of cores to use for parallel computation (default is 1).}
}
\value{
Tibble with columns:
\itemize{
  \item \code{ntop}: Number of top predicted edges evaluated.
  \item \code{nedge_true}: Number of edges in the true graph.
  \item \code{nedge}: Number of edges in the estimated graph.
  \item \code{intersection}: Number of overlap edges.
  \item \code{jaccard}: Jaccard index between the edges of the two graphs.
  \item \code{recall}: Recall of the estimated graph.
  \item \code{precision}: Precision of the estimated graph.
  \item \code{f1}: F1 score of the estimated graph.
  \item \code{shd}: Structural Hamming Distance (SHD), as computed by
    \code{pcalg::shd()}.
}
}
\description{
This function extracts the subgraph with top predicted edges and compares its
structure with the true graph.
}
