% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_perturbation_effect.R
\name{calc_perturbation_effect}
\alias{calc_perturbation_effect}
\title{Calculate Perturbation Effect}
\usage{
calc_perturbation_effect(Y, group, ncores, use_disk = NULL)
}
\arguments{
\item{Y}{Matrix of normalized scRNA-seq data of wild-type and perturbed cells.}

\item{group}{Named vector of cell label: 'WT' for wild-type cells, perturbed gene for perturbed cell.}

\item{ncores}{Number of CPUs to use.}

\item{use_disk}{Write huge data matrix to disk for computation. Determined automatically under default \code{NULL}.}
}
\value{
Tibble with columns:
\itemize{
  \item \code{ko}: Perturbed gene.
  \item \code{gene}: Affected gene.
  \item \code{diff}: Mean difference (Perturbed - WT).
  \item \code{cd}: Cohen's D (Normalized mean difference).
  \item \code{wilcox_pv}: Wilcoxon rank sum test P-value.
  \item \code{t_pv}: T-test P-value.
  \item \code{cor_pearson}: Pearson correlation in wild-type cells.
  \item \code{cor_spearman}: Spearman correlation in all cells.
  \item \code{adj_wilcox_pv}: BH-adjusted Wilcoxon rank sum test P-value.
  \item \code{adj_t_pv}: BH-adjusted T-test P-value.
  \item \code{adj_pv}: The larger of `adj_wilcox_pv` and `adj_wilcox_pv`.
}
}
\description{
Calculate differential expression statistics between wild-type cells and perturbed cells for each perturbation.
}
\examples{
# x -> y -> z
set.seed(123)
n <- 1e3
x0 <- rnorm(n)
x1 <- rnorm(n, mean = -1)
x <- c(x0, x1)
group <- c(rep('WT', n), rep('x', n))
y <- rnorm(2 * n, mean = x, sd = 1)
z <- rnorm(2 * n, mean = y, sd = 1)
data <- cbind(x = x, y = y, z = z)
stat <- calc_perturbation_effect(
  Y = data,
  group = group,
  ncores = 2
)
print(stat)
}
