% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_graph_with_gene_program.R
\name{collapse_graph_with_gene_program}
\alias{collapse_graph_with_gene_program}
\title{Collapse graph with a gene program node}
\usage{
collapse_graph_with_gene_program(graph, pname, pgenes, Y)
}
\arguments{
\item{graph}{igraph object containing the gene program node.}

\item{pname}{Name of the gene program node.}

\item{pgenes}{Character vector of gene program member genes.}

\item{Y}{scRNA-seq normalized expression matrix with columns matching graph
vertex names.}
}
\value{
igraph object.
}
\description{
Converts each incoming edge to the gene program node into incoming edges to
all member genes, then removes the gene program node.
}
