% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_prediction.R
\name{fit_expression_model}
\alias{fit_expression_model}
\title{Fit Gene Expression Model}
\usage{
fit_expression_model(
  Y,
  group,
  graph,
  ncores,
  method = c("lm", "lasso", "ridge")
)
}
\arguments{
\item{Y}{A numeric matrix of expression data (cells x genes).}

\item{group}{A character or factor vector indicating the group for each cell in Y.}

\item{graph}{An igraph object or the character string 'all' specifying the
regulatory structure.}

\item{ncores}{The number of cores for parallel computation.}

\item{method}{The regression method to use. One of 'lm', 'lasso', or 'ridge'.}
}
\value{
A numeric matrix `B` where rows are sources ('Intercept' and genes)
  and columns are targets (genes).
}
\description{
Fits a predictive model for each gene's expression based on a given graph
structure and regression method.
}
