% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{format_grnboost2}
\alias{format_grnboost2}
\title{Format GRNBoost2 Output}
\usage{
format_grnboost2(file, genes)
}
\arguments{
\item{file}{GRNBoost2 output.}

\item{genes}{Gene names.}
}
\value{
igraph object.
}
\description{
Reads GRNBoost2 output csv file and returns an igraph object.
}
