% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_edge_index}
\alias{get_edge_index}
\title{Get Edge Index of Adjacency matrix}
\usage{
get_edge_index(G)
}
\arguments{
\item{G}{Adjacency matrix.}
}
\value{
Edge index matrix.
}
\description{
Get edge index of adjacency matrix.
}
