% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_causalgrn.R
\name{infer_causalgrn}
\alias{infer_causalgrn}
\title{Direct edges using perturbation effect}
\usage{
infer_causalgrn(
  graph,
  stat,
  alpha,
  conservative = TRUE,
  max_order = 1,
  max_dist = Inf,
  evidence = 1
)
}
\arguments{
\item{graph}{Initial igraph object.}

\item{stat}{Perturbation effect from \code{\link{calc_perturbation_effect}}.}

\item{alpha}{Numeric representing DE Q-value threshold.}

\item{conservative}{Logical indicating whether to make conservative inference (Default is \code{TRUE}).}

\item{max_order}{Integer representing the maximum order for DE descendant inference. Can be 1 or 2. Default is 1.}

\item{max_dist}{Integer representing the maximum distance allowed for a perturbed gene to cause DE on another gene. Ignore for `max_order = 1`. Default is Inf.}

\item{evidence}{Integer representing number of evidences needed for second-order orientation. Ignore for `max_order = 1`. Default is 1.}
}
\value{
igraph object.
}
\description{
Infers causal GRN by directing edges based on differential expression caused by perturbations.
}
\examples{
# No hub genes
# x -> y -> z
set.seed(123)
n <- 1e3
x0 <- rnorm(n)
x1 <- rnorm(n, mean = -1)
x <- c(x0, x1)
group <- c(rep('WT', n), rep('x', n))
y <- rnorm(2 * n, mean = x, sd = 1)
z <- rnorm(2 * n, mean = y, sd = 1)
data <- cbind(x = x, y = y, z = z)
skel <- infer_skeleton_from_perturbed(
  Y = data,
  group = group,
  alpha = 1e-3,
  ncores = 2,
  max_order = 1
)
stat <- calc_perturbation_effect(
  Y = data,
  group = group,
  ncores = 2
)
cgrn <- infer_causalgrn(
  graph = skel$graph,
  stat = stat,
  alpha = 0.05,
  max_order = 2
)
plot(cgrn)
print(igraph::as_data_frame(cgrn))

}
