% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_causalgrn_with_gene_program.R
\name{infer_causalgrn_with_gene_program}
\alias{infer_causalgrn_with_gene_program}
\title{Direct edges using perturbation effect with a gene program node}
\usage{
infer_causalgrn_with_gene_program(
  graph,
  stat,
  alpha,
  pname,
  pgenes,
  conservative = TRUE,
  max_order = 1
)
}
\arguments{
\item{graph}{Initial igraph object.}

\item{stat}{Perturbation effect from \code{\link{calc_perturbation_effect}}.}

\item{alpha}{Numeric representing DE Q-value threshold.}

\item{pname}{Name of the gene program node.}

\item{pgenes}{Character vector of gene program member genes.}

\item{conservative}{Logical indicating whether to make conservative inference (Default is \code{TRUE}).}

\item{max_order}{Integer representing the maximum order for DE descendant inference. Can be 1 or 2. Default is 1.}
}
\value{
igraph object.
}
\description{
Infers causal GRN by directing edges based on differential expression caused
by perturbations. Program genes are always allowed as parents of
non-program genes.
}
