% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_causalgrn_with_gp.R
\name{infer_causalgrn_with_gp}
\alias{infer_causalgrn_with_gp}
\title{Direct edges using perturbation effect with gene program}
\usage{
infer_causalgrn_with_gp(
  graph,
  stat,
  alpha,
  pname,
  pgenes,
  allow_pgene_out,
  conservative = TRUE,
  max_order = 1,
  max_dist = Inf,
  evidence = 1
)
}
\arguments{
\item{graph}{Initial igraph object.}

\item{stat}{Perturbation effect from \code{\link{calc_perturbation_effect}}.}

\item{alpha}{Numeric representing DE Q-value threshold.}

\item{pname}{Gene program name.}

\item{pgenes}{Gene program genes.}

\item{allow_pgene_out}{Allow program genes as parents of non-program genes or not.}

\item{conservative}{Logical indicating whether to make conservative inference (Default is \code{TRUE}).}

\item{max_order}{Integer representing the maximum order for DE descendant inference. Can be 1 or 2. Default is 1.}

\item{max_dist}{Integer representing the maximum distance allowed for a perturbed gene to cause DE on another gene. Ignore for `max_order = 1`. Default is Inf.}

\item{evidence}{Integer representing number of evidences needed for second-order orientation. Ignore for `max_order = 1`. Default is 1.}
}
\value{
igraph object.
}
\description{
Infers causal GRN by directing edges based on differential expression caused by perturbations.
}
