% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_skeleton_with_gene_program.R
\name{infer_skeleton_with_gene_program}
\alias{infer_skeleton_with_gene_program}
\title{Infer graph skeleton with a gene program node}
\usage{
infer_skeleton_with_gene_program(
  count,
  Y,
  alpha,
  min_abspcor,
  pname,
  pgenes,
  ncores,
  G = NULL,
  max_order = 1,
  max_thr = 10,
  min_n1 = 1000,
  min_n2 = 200
)
}
\arguments{
\item{count}{scRNA-seq count matrix (cells x nodes, including gene program).}

\item{Y}{scRNA-seq normalized expression matrix (cells × nodes, including
gene program). For example, log1p(total UMI corrected count).}

\item{alpha}{Significance level for CI tests.}

\item{min_abspcor}{Minimum absolute value of partial correlation for kept edges.}

\item{pname}{Gene program node name in \code{Y}, such as \code{"PC1"}.}

\item{pgenes}{Gene program member genes in \code{Y}.}

\item{ncores}{Number of CPU cores for parallel processing.}

\item{G}{Optional initial adjacency matrix (defaults to a fully connected
graph without self-loops).}

\item{max_order}{Maximum conditioning set size (0 or 1, default is 1).}

\item{max_thr}{Maximum threshold for conditional variable (default is 10).}

\item{min_n1}{Minimum number of samples satisfying Yk > selected threshold
(default is 1000).}

\item{min_n2}{Minimum number of samples satisfying Yk > selected threshold,
Yi > 0, and Yj > 0 (default is 200).}
}
\value{
An igraph object representing the inferred skeleton.
}
\description{
Constructs a graph skeleton by performing iterative conditional independence
(CI) tests on all columns of the input matrices, including the gene program
node. Missing bidirectional skeleton edges are then added between the gene
program node and its member genes.
}
