% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_skeleton_with_gp.R
\name{infer_skeleton_with_gp}
\alias{infer_skeleton_with_gp}
\title{Infer graph skeleton with gene program}
\usage{
infer_skeleton_with_gp(
  count,
  Y,
  alpha,
  min_abspcor,
  pname,
  pgenes,
  ncores,
  G = NULL,
  max_order = 1,
  max_thr = 10,
  min_n1 = 2000,
  min_n2 = 400,
  max_nchildren = Inf,
  max_nparent = Inf,
  sepset = TRUE
)
}
\arguments{
\item{count}{scRNA-seq count matrix (cells x genes).}

\item{Y}{scRNA-seq normalized expression matrix (cells × genes, including gene program). For example, log1p(total UMI corrected count).}

\item{alpha}{Significance level for CI tests.}

\item{min_abspcor}{Minimum absolute value of partial correlation for kept edges.}

\item{pname}{Gene program name in \code{Y}.}

\item{pgenes}{Gene program genes in \code{Y}.}

\item{ncores}{Number of CPU cores for parallel processing.}

\item{G}{Optional initial adjacency matrix (defaults to a fully connected graph without self-loops).}

\item{max_order}{Maximum conditioning set size (0 or 1, default is 1).}

\item{max_thr}{Maximum threshold for conditional variable (default is 10).}

\item{min_n1}{Minimum number of samples satisfying Yk > selected threshold (default is 2000).}

\item{min_n2}{Minimum number of samples satisfying Yk > selected threshold, Yi > 0, and Yj > 0 (default is 400).}

\item{max_nchildren}{Maximum number of children a node can have (default is Inf).}

\item{max_nparent}{Maximum number of parents a node can have (default is Inf).}

\item{sepset}{Return separation set or not (default is \code{TRUE}).}
}
\value{
A list with:
\describe{
  \item{graph}{An igraph object representing the inferred skeleton with edge attributes `pMax`, `chisqMin`, and `absPcorMin`.}
  \item{sepSet}{List of separation sets for each node pair.}
}
}
\description{
Constructs graph skeleton by performing iterative conditional independence (CI) tests on gene expression data.
Edges are removed based on statistical independence.
}
