% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_perturbation_effect.R
\name{plot_cohens_d_vs_cor}
\alias{plot_cohens_d_vs_cor}
\title{Scatter plot of Correlation versus Cohens' D}
\usage{
plot_cohens_d_vs_cor(stat)
}
\arguments{
\item{stat}{Calculated from \code{\link{calc_perturbation_effect}}.}
}
\value{
ggplot object.
}
\description{
Draw scatter plot between Cohen's D and correlation in wild-type cells of KO-gene pairs.
}
\examples{
# x -> y -> z
set.seed(123)
n <- 1e3
x0 <- rnorm(n)
x1 <- rnorm(n, mean = -1)
x <- c(x0, x1)
group <- c(rep('WT', n), rep('x', n))
y <- rnorm(2 * n, mean = x, sd = 1)
z <- rnorm(2 * n, mean = y, sd = 1)
data <- cbind(x = x, y = y, z = z)
stat <- calc_perturbation_effect(
  Y = data,
  group = group,
  ncores = 2
)
plot_cohens_d_vs_cor(stat)

}
