% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_prediction_with_gp.R
\name{predict_standard_effect_with_gp}
\alias{predict_standard_effect_with_gp}
\title{Predict Perturbation Effect Using a Data-Driven SCOPE Imputation}
\usage{
predict_standard_effect_with_gp(
  B,
  scope_mean_effect,
  ko_expressions,
  wt_expressions,
  pname,
  pgenes,
  max_dist = Inf,
  ncores = 1
)
}
\arguments{
\item{B}{A numeric matrix of regulatory coefficients (sources x targets).}

\item{scope_mean_effect}{A named vector of the average training DE for the hub
and signature genes.}

\item{ko_expressions}{A named vector of expression levels for perturbed genes.}

\item{wt_expressions}{A named vector of WT expressions.}

\item{pname}{The name of the hub node (e.g., "PC1").}

\item{pgenes}{A character vector of the signature genes.}

\item{max_dist}{Maximum distance for effect propagation.}

\item{ncores}{Number of cores for parallel computation.}
}
\value{
A numeric matrix of predicted delta values (KOs x genes).
}
\description{
Predicts KO effects using a hybrid model. If a KO is predicted to affect the
SCOPE signature, the effect on the hub and signature genes is imputed using a
pre-calculated average effect. Effects on non-signature genes are then
calculated using a linear solver.
}
