% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{run_gies}
\alias{run_gies}
\title{Run GIES Algorithm}
\usage{
run_gies(wt, pts, verbose = FALSE)
}
\arguments{
\item{wt}{scRNA-seq matrix of wild-type cells.}

\item{pts}{List of scRNA-seq data matrices of perturbed cells, with names representing perturbed gene.}

\item{verbose}{Whether to print training details (default is FALSE).}
}
\value{
igraph object.
}
\description{
Run GIES algorithm and return an igraph object.
}
