% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{run_lasso}
\alias{run_lasso}
\title{Infer GRN with Lasso}
\usage{
run_lasso(Y, ncores, nfold = 5)
}
\arguments{
\item{Y}{Matrix of normalized scRNA-seq data.}

\item{ncores}{Number of cores to use for parallel computation.}

\item{nfold}{Number of folds for cross-validation (default is 5).}
}
\value{
igraph object.
}
\description{
Infers GRN with lasso, treating all cells as homogeneous.
}
