% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines_grn.R
\name{run_pc}
\alias{run_pc}
\title{Run PC Algorithm}
\usage{
run_pc(wt, alpha, ncores = 1)
}
\arguments{
\item{wt}{scRNA-seq matrix of wild-type cells.}

\item{alpha}{Significance level for conditional independence tests.}

\item{ncores}{Number of cores to use for parallel computation (default is 1).}
}
\value{
igraph object.
}
\description{
Runs PC algorithm and return an igraph object.
}
