% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{run_pc}
\alias{run_pc}
\title{Run PC Algorithm}
\usage{
run_pc(wt, alpha)
}
\arguments{
\item{wt}{scRNA-seq matrix of wild-type cells.}

\item{alpha}{Significance level for conditional independence tests.}
}
\value{
igraph object.
}
\description{
Runs PC algorithm and return an igraph object.
}
