% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_expression.R
\name{simulate_grn_guided_expression}
\alias{simulate_grn_guided_expression}
\title{GRN-guided simulation of scRNA-seq data of wild-type and perturbed cells mimicking real data}
\usage{
simulate_grn_guided_expression(
  d,
  Y,
  group,
  min_coef = 0.3,
  max_coef = 0.5,
  center_normal_sd = 1,
  center_ko_eff = 0.9
)
}
\arguments{
\item{d}{Expected number of neighbors (in/out).}

\item{Y}{Real single-cell perturbation data (cells x genes).}

\item{group}{Named vector of cell label: 'WT' for wild-type cells, perturbed gene for perturbed cell.}

\item{min_coef}{Minimum value of simulated coefficients.}

\item{max_coef}{Maximum value of simulated coefficients.}

\item{center_normal_sd}{Center of standard deviations of latent normal variables. Gene specific SDs are sampled from Gamma distribution.}

\item{center_ko_eff}{Center of KO effectiveness. KO specific effectiveness is sampled from Beta distribution.}
}
\value{
List containing:
\itemize{
  \item \code{dag}: Simulated DAG.
  \item \code{coef}: Simulated coefficients.
  \item \code{Y}: Simulated expression.
  \item \code{group}: Simulated group.
}
}
\description{
Simulates gene expression based on simulated directed acyclic graph and real single-cell perturbation data.
}
