% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_difference.R
\name{structure_difference}
\alias{structure_difference}
\title{Evaluate Structure Difference to True Graph}
\usage{
structure_difference(g, g0)
}
\arguments{
\item{g}{igraph object of the estimated graph.}

\item{g0}{igraph object of the true graph.}
}
\value{
Tibble with columns:
\itemize{
  \item \code{nedge_true}: Number of edges in the true graph.
  \item \code{nedge}: Number of edges in the estimated graph.
  \item \code{intersection}: Number of overlap edges.
  \item \code{jaccard}: Jaccard index between the edges of the two graphs.
  \item \code{recall}: Recall of the estimated graph.
  \item \code{precision}: Precision of the estimated graph.
  \item \code{f1}: F1 score of the estimated graph.
  \item \code{shd}: Structural Hamming Distance (SHD) between the true and estimated graphs.
}
}
\description{
This function evaluates the structure difference between the estimated directed graph and the true graph.
}
