% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_difference.R
\name{structure_difference_at_breaks}
\alias{structure_difference_at_breaks}
\title{Evaluate Structure Difference to True Graph at Different Breaks of Top Predicted Edges}
\usage{
structure_difference_at_breaks(pred, g0, ntops)
}
\arguments{
\item{pred}{Data frame of predicted directed edges with column \code{score} representing edge ranking.}

\item{g0}{igraph object of the true graph.}

\item{ntops}{Vector of number of top predicted edges to evaluate.}
}
\value{
Tibble with columns:
\itemize{
  \item \code{ntop}: Number of top predicted edges evaluated.
  \item \code{nedge_true}: Number of edges in the true graph.
  \item \code{nedge}: Number of edges in the estimated graph.
  \item \code{intersection}: Number of overlap edges.
  \item \code{jaccard}: Jaccard index between the edges of the two graphs.
  \item \code{recall}: Recall of the estimated graph.
  \item \code{precision}: Precision of the estimated graph.
  \item \code{f1}: F1 score of the estimated graph.
  \item \code{shd}: Structural Hamming Distance (SHD) between the true and estimated graphs.
}
}
\description{
This function extract the subgraph with top predicted edges and compare its structure with the true graph.
}
